hardened franchise delete no auth, sql injection update user, /api/auth requires password

+++ src/routes/franchiseRouter.js
@@ -96,7 +96,11 @@ franchiseRouter.post(
 // deleteFranchise
 franchiseRouter.delete(
   '/:franchiseId',
+  authRouter.authenticateToken,
   asyncHandler(async (req, res) => {
+    if(!req.user.isRole(Role.admin)){
+      throw new StatusCodeError('unable to delete store', 403)
+    }
     const franchiseId = Number(req.params.franchiseId);
     await DB.deleteFranchise(franchiseId);
     res.json({ message: 'franchise deleted' });
